STYLE: Replace `static_cast<unsigned int>` on integers with `U` suffixes
Following C++ Core Guidelines, Jun 14, 2026, Avoid casts, https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#res-casts

Uppercase suffix `U` appears (somewhat?) preferable to lowercase `u`, according to https://clang.llvm.org/extra/clang-tidy/checks/readability/uppercase-literal-suffix.html

+++ Components/Optimizers/CMAEvolutionStrategy/itkCMAEvolutionStrategyOptimizer.cxx
@@ -318,7 +318,7 @@ CMAEvolutionStrategyOptimizer::InitializeConstants()
   {
     this->m_UpdateBDPeriod = static_cast<unsigned int>(std::floor(1.0 / c_cov / Nd / 10.0));
   }
-  this->m_UpdateBDPeriod = std::max(static_cast<unsigned int>(1), this->m_UpdateBDPeriod);
+  this->m_UpdateBDPeriod = std::max(1U, this->m_UpdateBDPeriod);
   if (this->m_UpdateBDPeriod >= this->m_MaximumNumberOfIterations)
   {
     this->SetUseCovarianceMatrixAdaptation(false);
